workers/cloud/requirements.txt
# Modal Cloud Worker Dependencies

# Modal SDK for serverless deployment
modal==0.63.0

# Whisper for transcription
openai-whisper==20231117

# AWS SDK for S3 operations
boto3==1.34.0
botocore==1.34.0

# HTTP requests
requests==2.31.0

# FFmpeg Python bindings (optional, for advanced audio processing)
ffmpeg-python==0.2.0

# Development dependencies (for local testing)
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
